requirements.txt
typer[all]==0.12.5
rich==13.9.2
PyPDF2==3.0.1
PyYAML==6.0.2
pytest==8.3.3
pdfminer.six==20240706
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.2
faiss-cpu==1.8.0
sentence-transformers==3.2.1
rank_bm25==0.2.2
tiktoken==0.7.0
fastapi==0.115.5
uvicorn==0.32.1
openai==1.56.0
langgraph==0.2.34
python-dotenv==1.0.1
httpx==0.27.0
pylint==4.0.4
black==25.12.0
